+++ lib/twilio/device.ts
@@ -497,13 +497,11 @@ class Device extends EventEmitter {
         Twilio Support at <help@twilio.com>.`);
     }
 
-    if (window) {
-      const root: any = window as any;
-      const browser: any = root.msBrowser || root.browser || root.chrome;
+    const root: any = globalThis as any;
+    const browser: any = root.msBrowser || root.browser || root.chrome;
 
-      this._isBrowserExtension = (!!browser && !!browser.runtime && !!browser.runtime.id)
-        || (!!root.safari && !!root.safari.extension);
-    }
+    this._isBrowserExtension = (!!browser && !!browser.runtime && !!browser.runtime.id)
+      || (!!root.safari && !!root.safari.extension);
 
     if (this._isBrowserExtension) {
       this._log.info('Running as browser extension.');
@@ -559,22 +557,51 @@ class Device extends EventEmitter {
    * @param options
    */
   async connect(options: Device.ConnectOptions = { }): Promise<Call> {
-    this._log.debug('.connect', JSON.stringify(options && options.params || {}), options);
+    this._log.debug('.connect', JSON.stringify(options));
     this._throwIfDestroyed();
-
     if (this._activeCall) {
       throw new InvalidStateError('A Call is already active');
     }
 
-    const activeCall = this._activeCall = await this._makeCall(
-      options.params || { },
-      {
-        enableImprovedSignalingErrorPrecision:
-          !!this._options.enableImprovedSignalingErrorPrecision,
-        rtcConfiguration: options.rtcConfiguration,
-        voiceEventSidGenerator: this._options.voiceEventSidGenerator,
-      },
-    );
+    let customParameters;
+    let parameters;
+    let signalingReconnectToken;
+
+    if (options.connectToken) {
+      try {
+        const connectTokenParts = JSON.parse(decodeURIComponent(atob(options.connectToken)));
+        customParameters = connectTokenParts.customParameters;
+        parameters = connectTokenParts.parameters;
+        signalingReconnectToken = connectTokenParts.signalingReconnectToken;
+      } catch {
+        throw new InvalidArgumentError('Cannot parse connectToken');
+      }
+
+      if (!parameters || !parameters.CallSid || !signalingReconnectToken) {
+        throw new InvalidArgumentError('Invalid connectToken');
+      }
+    }
+
+    let isReconnect = false;
+    let twimlParams: Record<string, string> = {};
+    const callOptions: Call.Options = {
+      enableImprovedSignalingErrorPrecision:
+        !!this._options.enableImprovedSignalingErrorPrecision,
+      rtcConfiguration: options.rtcConfiguration,
+      voiceEventSidGenerator: this._options.voiceEventSidGenerator,
+    };
+
+    if (signalingReconnectToken && parameters) {
+      isReconnect = true;
+      callOptions.callParameters = parameters;
+      callOptions.reconnectCallSid = parameters.CallSid;
+      callOptions.reconnectToken = signalingReconnectToken;
+      twimlParams = customParameters || twimlParams;
+    } else {
+      twimlParams = options.params || twimlParams;
+    }
+
+    const activeCall = this._activeCall = await this._makeCall(twimlParams, callOptions, isReconnect);
 
     // Make sure any incoming calls are ignored
     this._calls.splice(0).forEach(call => call.ignore());
@@ -987,7 +1014,7 @@ class Device extends EventEmitter {
    * @param twimlParams - A flat object containing key:value pairs to be sent to the TwiML app.
    * @param options - Options to be used to instantiate the {@link Call}.
    */
-  private async _makeCall(twimlParams: Record<string, string>, options?: Call.Options): Promise<Call> {
+  private async _makeCall(twimlParams: Record<string, string>, options?: Call.Options, isReconnect: boolean = false): Promise<Call> {
     if (typeof Device._isUnifiedPlanDefault === 'undefined') {
       throw new InvalidStateError('Device has not been initialized.');
     }
@@ -1056,7 +1083,7 @@ class Device extends EventEmitter {
         this._audio._maybeStartPollingVolume();
       }
 
-      if (call.direction === Call.CallDirection.Outgoing && this._audio?.outgoing()) {
+      if (call.direction === Call.CallDirection.Outgoing && this._audio?.outgoing() && !isReconnect) {
         this._soundcache.get(Device.SoundName.Outgoing).play();
       }
 
@@ -1645,9 +1672,40 @@ namespace Device {
   declare function errorEvent(error: TwilioError, call?: Call): void;
 
   /**
-   * Emitted when an incoming {@link Call} is received.
+   * Emitted when an incoming {@link Call} is received. You can interact with the call object
+   * using its public APIs, or you can forward it to a different {@link Device} using
+   * {@link Device.connect} and {@link Call.connectToken}, enabling your application to
+   * receive multiple incoming calls for the same identity.
+   *
+   * **Important:** When forwarding a call, the token for target device instance needs to have
+   * the same identity as the token used in the device that originally received the call.
+   *
+   * @example
+   *
+   * ```js
+   * const receiverDevice = new Device(token, options);
+   * await receiverDevice.register();
+   *
+   * receiverDevice.on('incoming', (call) => {
+   *   // Forward this call to a new Device instance using the call.connectToken string.
+   *   forwardCall(call.connectToken);
+   * });
+   *
+   * // The forwardCall function may look something like the following.
+   * async function forwardCall(connectToken) {
+   *   // For every incoming call, we create a new Device instance which we can
+   *   // interact with, without affecting other calls.
+   *   // IMPORTANT: The token for this new device needs to have the same identity
+   *   // as the token used in the receiverDevice.
+   *   const device = new Device(token, options);
+   *   const call = await device.connect({ connectToken });
+   *
+   *   // Destroy the device after the call is completed
+   *   call.on('disconnect', () => device.destroy());
+   * }
+   * ```
+   *
    * @param call - The incoming {@link Call}.
-   * @example `device.on('incoming', call => { })`
    * @event
    */
   declare function incomingEvent(call: Call): void;
@@ -1738,6 +1796,20 @@ namespace Device {
    * Options to be passed to {@link Device.connect}.
    */
   export interface ConnectOptions extends Call.AcceptOptions {
+    /**
+     * The {@link Call.connectToken} to use to manually reconnect to an existing call.
+     * A call can be manually reconnected if it was previously received (incoming)
+     * or created (outgoing) from a {@link Device} instance.
+     * A call is considered manually reconnected if it was created using the {@link Call.connectToken}.
+     * It will always have a {@link Call.direction} property set to {@link Call.CallDirection.Outgoing}.
+     *
+     * **Warning: Only unanswered incoming calls can be manually reconnected at this time.**
+     * **Invoking this method to an already answered call may introduce unexpected behavior.**
+     *
+     * See {@link Device.incomingEvent} for an example.
+     */
+    connectToken?: string;
+
     /**
      * A flat object containing key:value pairs to be sent to the TwiML app.
      */